Extract MCP RMCP client handling
Move the managed RMCP client lifecycle, startup, construction, and uncached tool listing into codex-mcp/src/client.rs. Keep cache and shared connection data in mcp_connection.rs, and update manager imports/tests for the new module boundary.
