+++ codex-rs/codex-mcp/src/manager.rs
@@ -5,22 +5,22 @@ use std::time::Duration;
 use std::time::Instant;
 
 use crate::McpAuthStatusEntry;
+use crate::client::AsyncManagedClient;
+use crate::client::MCP_TOOLS_FETCH_UNCACHED_DURATION_METRIC;
+use crate::client::MCP_TOOLS_LIST_DURATION_METRIC;
+use crate::client::ManagedClient;
+use crate::client::StartupOutcomeError;
+use crate::client::list_tools_for_client_uncached;
 use crate::mcp::CODEX_APPS_MCP_SERVER_NAME;
 use crate::mcp::ToolPluginProvenance;
-use crate::mcp_connection::AsyncManagedClient;
 use crate::mcp_connection::CodexAppsToolsCacheContext;
 use crate::mcp_connection::CodexAppsToolsCacheKey;
 use crate::mcp_connection::ElicitationRequestManager;
-use crate::mcp_connection::MCP_TOOLS_FETCH_UNCACHED_DURATION_METRIC;
-use crate::mcp_connection::MCP_TOOLS_LIST_DURATION_METRIC;
-use crate::mcp_connection::ManagedClient;
 use crate::mcp_connection::McpRuntimeEnvironment;
-use crate::mcp_connection::StartupOutcomeError;
 use crate::mcp_connection::ToolInfo;
 use crate::mcp_connection::emit_duration;
 use crate::mcp_connection::emit_update;
 use crate::mcp_connection::filter_tools;
-use crate::mcp_connection::list_tools_for_client_uncached;
 use crate::mcp_connection::mcp_init_error_display;
 use crate::mcp_connection::qualify_tools;
 use crate::mcp_connection::startup_outcome_error_message;
